+++ oshi-core/src/main/java/oshi/software/os/FileSystem.java
@@ -61,4 +61,12 @@ public interface FileSystem {
      * @return The maximum number of file descriptors if available, 0 otherwise.
      */
     long getMaxFileDescriptors();
+
+    /**
+     * The maximum number of open file descriptors per process. This returns the upper limit which applies to each
+     * process. The actual limit of a process may be lower if configured.
+     *
+     * @return The maximum number of file descriptors of each process if available, 0 otherwise.
+     */
+    long getMaxFileDescriptorsPerProcess();
 }